pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vecorel-cli"
version = "0.1.0"
description = "CLI tools such as validation and file format conversion for vecorel."
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Matthias Mohr"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
# This is duplicate, must also be changed below in the pixi configuration
dependencies = [
    "pyyaml>=6.0,<7.0",
    "click>=8.1,<9.0",
    "geopandas>=1.0.0,<2.0",
    "requests>=2.30,<3.0",
    "shapely>=2.1,<3.0",
    # numpy is restricted <2.2.0 due to the lower supported Python version being 3.11 and we still cater for 3.10
    "numpy>=2.0,<2.2",
    "pyarrow>=21.0,<22.0",
    "py7zr>=1.0,<2.0",
    "fsspec==2025.7.0",
    "jsonschema[format]>=4.20,<5.0",
    "aiohttp>=3.9,<4.0",
    "yarl>=1.20,<2.0",
    "rarfile>=4.0,<5.0",
    "semantic-version>=2.10.0,<3.0",
    "json-stream>=2.3.0,<3.0",
    "flatdict>=4.0,<5.0",
    "loguru==0.7.3",
]

[project.scripts]
vec = "vecorel_cli:vecorel_cli"

[project.urls]
"Homepage" = "https://github.com/vecorel/cli"
"Bug Reports" = "https://github.com/vecorel/cli/issues"
"Source" = "https://github.com/vecorel/cli"

[tool.setuptools.packages.find]
include = ["vecorel_cli*"]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["win-64", "linux-64", "osx-64", "osx-arm64"]

[tool.pixi.dependencies]
python = ">=3.10"
pyyaml = ">=6.0,<7.0"
click = ">=8.1,<9.0"
geopandas = ">=1.0.0,<2.0"
requests = ">=2.30,<3.0"
shapely = ">=2.1,<3.0"
# numpy is restricted <2.2.0 due to the lower supported Python version being 3.11 and we still cater for 3.10
numpy = ">=2.0,<2.2"
pyarrow = ">=21.0,<22.0"
py7zr = ">=1.0,<2.0"

[tool.pixi.pypi-dependencies]
# Editable install of the project itself
vecorel-cli = {path = ".", editable = true}

[tool.pixi.feature.dev.dependencies]
pytest = ">=8.4.1,<9.0.0"
pytest-cov = ">=6.0.0,<7.0.0"
pre-commit = ">=4.2.0,<5.0.0"
ruff = "==0.12.8"
GitPython = ">=3.0.0,<4.0.0"

[tool.pixi.feature.dev.pypi-dependencies]
build = ">=1.0.0,<2.0.0"

[tool.pixi.feature.cloud.dependencies]
s3fs = "==2025.7.0"
gcsfs = "==2025.7.0"

[tool.pixi.environments]
default = {solve-group = "default"}
dev = {features = ["dev"], solve-group = "default"}
cloud = {features = ["cloud"], solve-group = "default"}
all = {features = ["dev", "cloud"], solve-group = "default"}

[tool.pixi.tasks]
# Development tasks
install-dev = "pip install -e ."
test = "pytest --cov=vecorel_cli --cov-report=term-missing --cov-report=html --cov-report=xml"
lint = "ruff check --fix"
format = "ruff format"
check = {depends-on = ["lint", "format", "test"]}
build = "python -m build"

# Pre-commit tasks
pre-commit-install = "pre-commit install"
pre-commit-run = "pre-commit run --all-files"

# CLI entry point
vec = "vec"

[tool.ruff]
exclude = [
    ".eggs",
    ".git",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "build",
    "dist",
    "site-packages",
    "tests/datasets/invalid_*.py",
]

line-length = 100
indent-width = 4

target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "W", "I", "NPY201"]  # "C901"
ignore = []

fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.coverage.run]
source = ["vecorel_cli"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/.*",
]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
precision = 2
fail_under = 64
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"
